+++ ibc/utils_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"github.com/tharsis/evmos/v3/x/claims/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -56,6 +57,7 @@ func TestGetTransferSenderRecipient(t *testing.T) {
 					&transfertypes.FungibleTokenPacketData{
 						Sender:   "cosmos1",
 						Receiver: "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
+						Amount:   "123456",
 					},
 				),
 			},
@@ -69,6 +71,7 @@ func TestGetTransferSenderRecipient(t *testing.T) {
 					&transfertypes.FungibleTokenPacketData{
 						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
 						Receiver: "evmos1",
+						Amount:   "123456",
 					},
 				),
 			},
@@ -82,6 +85,7 @@ func TestGetTransferSenderRecipient(t *testing.T) {
 					&transfertypes.FungibleTokenPacketData{
 						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
 						Receiver: "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
+						Amount:   "123456",
 					},
 				),
 			},
@@ -96,6 +100,7 @@ func TestGetTransferSenderRecipient(t *testing.T) {
 					&transfertypes.FungibleTokenPacketData{
 						Sender:   "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
 						Receiver: "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
+						Amount:   "123456",
 					},
 				),
 			},
@@ -110,6 +115,7 @@ func TestGetTransferSenderRecipient(t *testing.T) {
 					&transfertypes.FungibleTokenPacketData{
 						Sender:   "osmo1qql8ag4cluz6r4dz28p3w00dnc9w8ueuhnecd2",
 						Receiver: "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
+						Amount:   "123456",
 					},
 				),
 			},
@@ -127,6 +133,96 @@ func TestGetTransferSenderRecipient(t *testing.T) {
 			require.NoError(t, err, tc.name)
 			require.Equal(t, tc.expSender, sender.String())
 			require.Equal(t, tc.expRecipient, recipient.String())
+		}
+	}
+}
+
+func TestGetTransferAmount(t *testing.T) {
+	testCases := []struct {
+		name      string
+		packet    channeltypes.Packet
+		expAmount string
+		expError  bool
+	}{
+		{
+			"empty packet",
+			channeltypes.Packet{},
+			"",
+			true,
+		},
+		{
+			"invalid packet data",
+			channeltypes.Packet{
+				Data: ibctesting.MockFailPacketData,
+			},
+			"",
+			true,
+		},
+		{
+			"invalid amount - empty",
+			channeltypes.Packet{
+				Data: transfertypes.ModuleCdc.MustMarshalJSON(
+					&transfertypes.FungibleTokenPacketData{
+						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
+						Receiver: "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
+						Amount:   "",
+					},
+				),
+			},
+			"",
+			true,
+		},
+		{
+			"invalid amount - non-int",
+			channeltypes.Packet{
+				Data: transfertypes.ModuleCdc.MustMarshalJSON(
+					&transfertypes.FungibleTokenPacketData{
+						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
+						Receiver: "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
+						Amount:   "test",
+					},
+				),
+			},
+			"test",
+			true,
+		},
+		{
+			"valid",
+			channeltypes.Packet{
+				Data: transfertypes.ModuleCdc.MustMarshalJSON(
+					&transfertypes.FungibleTokenPacketData{
+						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
+						Receiver: "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
+						Amount:   "10000",
+					},
+				),
+			},
+			"10000",
+			false,
+		},
+		{
+			"valid - IBCTriggerAmt",
+			channeltypes.Packet{
+				Data: transfertypes.ModuleCdc.MustMarshalJSON(
+					&transfertypes.FungibleTokenPacketData{
+						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
+						Receiver: "evmos1x2w87cvt5mqjncav4lxy8yfreynn273xn5335v",
+						Amount:   types.IBCTriggerAmt,
+					},
+				),
+			},
+			types.IBCTriggerAmt,
+			false,
+		},
+	}
+
+	for _, tc := range testCases {
+		amt, err := GetTransferAmount(tc.packet)
+		if tc.expError {
+			require.Error(t, err, tc.name)
+		} else {
+			require.NoError(t, err, tc.name)
+			require.Equal(t, tc.expAmount, amt)
 		}
 	}
 }