+++ x/claims/keeper/ibc_callbacks.go
@@ -114,16 +114,16 @@ func (k Keeper) OnRecvPacket(
 	// supports ethereum keys (eg: Cronos, Injective).
 	if sameAddress && !fromEVMChain {
 		switch {
-		// case 1: secp256k1 key from sender/recipient has no claimed actions
-		// -> return error acknowledgement to prevent funds from getting stuck
 		case senderRecordFound && !senderClaimsRecord.HasClaimedAny():
+			// secp256k1 key from sender/recipient has no claimed actions
+			// -> return error acknowledgement to prevent funds from getting stuck
 			return channeltypes.NewErrorAcknowledgement(
 				sdkerrors.Wrapf(
 					evmos.ErrKeyTypeNotSupported, "receiver address %s is not a valid ethereum address", recipientBech32,
 				).Error(),
 			)
-		// case 2: sender/recipient has funds stuck -> return ack to trigger withdrawal
 		default:
+			// sender/recipient has funds stuck -> return ack to trigger withdrawal
 			return ack
 		}
 	}
@@ -140,11 +140,18 @@ func (k Keeper) OnRecvPacket(
 	}
 
 	recipientClaimsRecord, recipientRecordFound := k.GetClaimsRecord(ctx, recipient)
+	amt, err := ibc.GetTransferAmount(packet)
+	if err != nil {
+		return channeltypes.NewErrorAcknowledgement(err.Error())
+	}
+	isTriggerAmt := amt == types.IBCTriggerAmt
 
-	// Handle four cases of comparing recipient and sender addresses and checking
-	// for claims records
 	switch {
-	case senderRecordFound && recipientRecordFound && !sameAddress:
+	// Cases with SenderRecordFound.
+	// They require a merge or migration of claims records. To prevent this
+	// happening by accident, they are only executed, when the sender transfers
+	// the specified IBCTriggerAmt.
+	case senderRecordFound && recipientRecordFound && !sameAddress && isTriggerAmt:
 		// case 1: both sender and recipient are distinct and have a claims record
 		// -> merge sender's record with the recipient's record and claim actions that
 		// have already been claimed by one or the other
@@ -163,8 +170,7 @@ func (k Keeper) OnRecvPacket(
 			"receiver", recipientBech32,
 			"total-claimable", senderClaimsRecord.InitialClaimableAmount.Add(recipientClaimsRecord.InitialClaimableAmount).String(),
 		)
-
-	case senderRecordFound && !recipientRecordFound:
+	case senderRecordFound && !recipientRecordFound && isTriggerAmt:
 		// case 2: only the sender has a claims record
 		// -> migrate the sender record to the recipient address and claim IBC action
 		k.SetClaimsRecord(ctx, recipient, senderClaimsRecord)
@@ -179,12 +185,11 @@ func (k Keeper) OnRecvPacket(
 
 		_, err = k.ClaimCoinsForAction(ctx, recipient, senderClaimsRecord, types.ActionIBCTransfer, params)
 
+	// Cases without SenderRecordFound
 	case !senderRecordFound && recipientRecordFound,
-		// case 3: only the recipient has a claims record
-		// -> only claim IBC transfer action
 		sameAddress && fromEVMChain && recipientRecordFound:
+		// case 3: only the recipient has a claims record -> only claim IBC transfer action
 		_, err = k.ClaimCoinsForAction(ctx, recipient, recipientClaimsRecord, types.ActionIBCTransfer, params)
-
 	case !senderRecordFound && !recipientRecordFound:
 		// case 4: neither the sender or recipient have a claims record
 		// -> perform a no-op by returning the original success acknowledgement